feat: refactor wallet types
