+++ src/renderer/domain/shared-kernel.ts
@@ -40,16 +40,3 @@ export const enum ErrorType {
   PATTERN = 'pattern',
   MAX_LENGTH = 'maxLength',
 }
-
-export const enum WalletType {
-  WATCH_ONLY = 'wallet_wo',
-  SINGLE_PARITY_SIGNER = 'wallet_sps',
-  MULTISHARD_PARITY_SIGNER = 'wallet_mps',
-  MULTISIG = 'wallet_ms',
-}
-
-export const enum SigningType {
-  WATCH_ONLY = 'signing_wo',
-  PARITY_SIGNER = 'signing_ps',
-  MULTISIG = 'signing_ms',
-}
