+++ src/renderer/pages/Onboarding/Vault/ManageStepSingle/ManageStepSingle.tsx
@@ -5,11 +5,12 @@ import { u8aToHex } from '@polkadot/util';
 
 import { useI18n } from '@renderer/app/providers';
 import { Chain } from '@renderer/entities/chain';
-import { ErrorType, SigningType } from '@renderer/domain/shared-kernel';
+import { ErrorType } from '@renderer/domain/shared-kernel';
 import { chainsService } from '@renderer/entities/network';
 import { Button, Input, InputHint, HeaderTitleText, SmallTitleText } from '@renderer/shared/ui';
 import { SeedInfo } from '@renderer/components/common/QrCode/common/types';
 import { useAccount, createAccount, AccountsList } from '@renderer/entities/account';
+import { SigningType } from '@renderer/entities/wallet';
 
 type WalletForm = {
   walletName: string;
