+++ src/renderer/pages/Staking/Overview/components/NominatorsList/NominatorsList.tsx
@@ -1,7 +1,7 @@
 import { ApiPromise } from '@polkadot/api';
 import { Trans } from 'react-i18next';
 
-import { Address, SigningType, EraIndex } from '@renderer/domain/shared-kernel';
+import { Address, EraIndex } from '@renderer/domain/shared-kernel';
 import { Unlocking } from '@renderer/entities/staking';
 import { useI18n } from '@renderer/app/providers';
 import { FootnoteText, Plate, Checkbox, InfoPopover, Tooltip, Icon, Shimmering, HelpText } from '@renderer/shared/ui';
@@ -11,6 +11,7 @@ import { Asset, AssetBalance } from '@renderer/entities/asset';
 import { TimeToEra } from '../TimeToEra/TimeToEra';
 import { redeemableAmount } from '@renderer/shared/lib/utils';
 import { AccountAddress } from '@renderer/entities/account';
+import { SigningType } from '@renderer/entities/wallet';
 
 const getNextUnstakingEra = (unlocking: Unlocking[] = [], era?: number): EraIndex | undefined => {
   if (!era) return undefined;
