+++ src/renderer/shared/lib/utils/constants.ts
@@ -1,5 +1,3 @@
-import { SigningType } from '@renderer/domain/shared-kernel';
-
 export const ZERO_BALANCE = '0';
 
 export const DEFAULT_TRANSITION = 300;
@@ -20,12 +18,6 @@ export const TEST_CHAIN_ID = '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb4
 export const TEST_HASH = '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3';
 export const TEST_CHAIN_ICON =
   'https://raw.githubusercontent.com/nova-wallet/nova-spektr-utils/main/icons/v1/assets/white/Polkadot_(DOT).svg';
-
-export const SigningBadges = {
-  [SigningType.PARITY_SIGNER]: 'paritySignerBg',
-  [SigningType.MULTISIG]: 'multisigBg',
-  [SigningType.WATCH_ONLY]: 'watchOnlyBg',
-} as const;
 
 export const enum KeyboardKey {
   ENTER = 'Enter',
