+++ src/renderer/shared/ui/Identicon/Identicon.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { SigningType } from '@renderer/domain/shared-kernel';
 import { TEST_ADDRESS } from '@renderer/shared/lib/utils';
+import { SigningType } from '@renderer/entities/wallet';
 import Identicon from './Identicon';
 
 export default {
