+++ src/renderer/widgets/CreateWallet/ui/CreateWalletProvider.tsx
@@ -2,11 +2,11 @@ import { useEffect } from 'react';
 import { useUnit } from 'effector-react';
 import { useNavigate } from 'react-router-dom';
 
-import { WalletType } from '@renderer/domain/shared-kernel';
 import { walletProviderModel } from '../model/wallet-provider-model';
 import WatchOnly from '@renderer/pages/Onboarding/WatchOnly/WatchOnly';
 import Vault from '@renderer/pages/Onboarding/Vault/Vault';
 import { MultisigAccount } from './MultisigAccount/MultisigAccount';
+import { WalletType } from '@renderer/entities/wallet';
 import { Paths } from '../../../app/providers/routes/paths';
 
 // TODO: Break down WatchOnly / Vault / CreateMultisig to widgets
